renaming the libs directory to site-packages and turning it into an actual site-packages directory to boot! the packages.pth is the package listing for the directory. `ls -1 | grep -v packages.pth > packages.pth` to update.
manage.py now uses the site.addsitedir() call to load the directory.
the paths loaded are then forcable prepended to sys.path as
addsitedir does an append (oh well).

This will allow us to have multiple projects with the same site-packages.
We might want to move common apps out to the apps directory and treat that
as a site-packages style setup as well? Then again, maybe not.